requirements.txt
discord.py>=2.3
python-dotenv>=1.0
PyJWT[crypto]>=2.8
httpx>=0.27
google-api-python-client>=2.100
google-auth>=2.23
google-cloud-storage>=2.14
fastapi>=0.115
uvicorn[standard]>=0.32
cryptography>=42.0
anthropic>=0.39
